+++ src/utility.rs
@@ -1,7 +1,7 @@
 pub use rand::*;
-pub use std::f64::consts::PI;
+pub use std::f32::consts::PI;
 
-pub fn clamp(x: f64, range: std::ops::Range<f64>) -> f64 {
+pub fn clamp(x: f32, range: std::ops::Range<f32>) -> f32 {
     if range.contains(&x) {
         x
     } else if range.start > x {
